refactor: extract mod tokenizer
